pyproject.toml
[project]
name = "e2b-gpt-engineer"
version = "0.1.0"
description = "A simple version of GPT Engineer using E2B sandboxes"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "e2b>=0.12.3",
    "fastapi>=0.100.0", 
    "uvicorn>=0.23.0",
    "python-dotenv>=1.0.0",
    "openai>=1.3.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "pydantic>=2.3.0",
]
requires-python = ">=3.9"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]